refactor(avatar): split into api and impl module
